Simplified Multi-Threading
I wrote this early on in my Python career and have constantly been finding myself reusing this code in projects.  You pass the threader some arguments on how to do the job and it does it, simple as that.
How many threads do you want?
What function are the jobs going to be mapped to?
What jobs should be passed to the mapped function by the enslaved threads?
The sample function I included with some sample usage, shows a problem that should normally take 20 seconds to complete since the function has time.sleep(1)...  Using 5 threads however, the program runs in only 4 seconds since the load is divided and executed concurrently by the called threads.